test(refactor): share RecordingProgress spy across tests #940
Both `NoEnvOutsideConfigHandlerTest` and the new
`PluginNoEnvFallbackTest` declared their own near-identical
`RecordingProgress` as a second top-level class in the test file.

Extract a single `Tests\Psalm\LaravelPlugin\Unit\Support\RecordingProgress`
spy and import it from both call sites. Drops 78 lines of duplication
and removes the two-classes-per-file pattern flagged in review.
