+++ tests/Unit/PluginNoEnvFallbackTest.php
@@ -18,6 +18,7 @@ use Psalm\LaravelPlugin\Providers\ApplicationProvider;
 use Psalm\Plugin\EventHandler\Event\FunctionReturnTypeProviderEvent;
 use Psalm\Progress\Progress;
 use Psalm\StatementsSource;
+use Tests\Psalm\LaravelPlugin\Unit\Support\RecordingProgress;
 
 /**
  * Plugin-level coverage for the implicit `<configDirectory>` fallback.
@@ -190,45 +191,5 @@ final class PluginNoEnvFallbackTest extends TestCase
         }
 
         \rmdir($path);
-    }
-}
-
-/**
- * Test-only Progress that records warnings without writing to STDERR.
- * Mirrors the sibling helper in NoEnvOutsideConfigHandlerTest; kept local so
- * the two test files stay independent (different namespaces; no collision).
- */
-final class RecordingProgress extends Progress
-{
-    public int $warningCount = 0;
-
-    public string $lastWarning = '';
-
-    #[\Override]
-    public function debug(string $message): void {}
-
-    #[\Override]
-    public function startPhase(\Psalm\Progress\Phase $phase, int $threads = 1): void {}
-
-    #[\Override]
-    public function expand(int $number_of_tasks): void {}
-
-    #[\Override]
-    public function taskDone(int $level): void {}
-
-    #[\Override]
-    public function finish(): void {}
-
-    #[\Override]
-    public function alterFileDone(string $file_name): void {}
-
-    #[\Override]
-    public function write(string $message): void {}
-
-    #[\Override]
-    public function warning(string $message): void
-    {
-        $this->warningCount++;
-        $this->lastWarning = $message;
     }
 }
